requirements.txt
# Kazakh Legal RAG Experiment Dependencies

# API clients
openai>=1.0.0
httpx>=0.25.0

# Vector store and search
faiss-cpu>=1.7.4
rank_bm25>=0.2.2

# Tokenization
tiktoken>=0.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Kazakh NLP (if available via pip, otherwise manual install)
# kaznlp  # Install from: https://github.com/nlacslab/kaznlp
